Move code for DMF self-testing (and sample usage) into separate file
